Removed redundant class, and created Filter interface to verify schedules based on specification use case classes.
