+++ src/main/java/Filter.java
@@ -1,0 +1,7 @@
+/**
+ * An interface that will be responsible for checking schedules
+ * based on given specifications.
+ */
+public interface Filter {
+    public boolean checkSchedule(Schedule s);
+}
